shutdown clogged connections
A clogged connection is when send() blocks the thread indefinitely.
With MSG_DONTWAIT flag, send() returns with errno set to EAGAIN if the
operation would block. It could mean that the connection is clogged or
the send buffer is full because of high traffic. We are only dealing
with the first option.